paru 1.1.2 (backport to gcc 7.5.0), and GraphBLAS 10.4.1

+++ GraphBLAS/Source/mxm/template/GB_AxB_dot3_meta.c
@@ -147,7 +147,7 @@
     #endif
 
     GB_Mi_DECLARE (Mi, const) ; GB_Mi_PTR (Mi, M) ;
-    const size_t mvlen = M->vlen ;
+    const int64_t mvlen = M->vlen ;
     const GB_M_TYPE *restrict Mx = (GB_M_TYPE *) (Mask_struct ? NULL : (M->x)) ;
 
     //--------------------------------------------------------------------------
